// spikes/s1-codegen/generated/PL_010f_v1.04/nfe.ts
export type TNFe_infNFe_emit = {
  /**
   * Razão Social ou Nome do emitente
   * xsd:TString, length 2..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  xNome: string;
  /**
   * Nome fantasia
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  xFant?: string;
  /** Endereço do emitente */
  enderEmit: TEnderEmi;
  /**
   * Inscrição Estadual do Emitente
   * xsd:TIe, length 0..14, pattern `[0-9]{2,14}|ISENTO`
   */
  IE?: string;
  /**
   * Inscricao Estadual do Substituto Tributário
   * xsd:TIeST, length 0..14, pattern `[0-9]{2,14}`
   */
  IEST?: string;
  /**
   * Inscrição Municipal
   * xsd:TString, length 1..15, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  IM?: string;
  /**
   * CNAE Fiscal
   * pattern `[0-9]{7}`
   */
  CNAE?: string;
  /**
   * Código de Regime Tributário.
   * Este campo será obrigatoriamente preenchido com:
   * 1 – Simples Nacional;
   * 2 – Simples Nacional – excesso de sublimite de receita bruta;
   * 3 – Regime Normal.
   * 4 - Simples Nacional - Microempreendedor individual - MEI
   */
  CRT: "1" | "2" | "3" | "4";
  /**
   * Inscrição do emitente na Suframa
   * pattern `[0-9]{8,9}`
   */
  ISUFEmit?: string;
} & (
  ({
  /**
   * Número do CNPJ do emitente
   * xsd:TCnpj, length 0..14, pattern `[0-9A-Z]{12}[0-9]{2}`
   */
  CNPJ: string;
  CPF?: never;
})
  | ({
  /**
   * Número do CPF do emitente
   * xsd:TCpf, length 0..11, pattern `[0-9]{11}`
   */
  CPF: string;
  CNPJ?: never;
})
);

/**
 * Emissão de avulsa, informar os dados do Fisco emitente
 * xsd: anonymous type of TNFe.infNFe.avulsa
 */
export type TNFe_infNFe_avulsa = {
  /**
   * CNPJ do Órgão emissor
   * xsd:TCnpj, length 0..14, pattern `[0-9A-Z]{12}[0-9]{2}`
   */
  CNPJ: string;
  /**
   * Órgão emitente
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  xOrgao: string;
  /**
   * Matrícula do agente
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  matr: string;
  /**
   * Nome do agente
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  xAgente: string;
  /**
   * Telefone
   * pattern `[0-9]{6,14}`
   */
  fone?: string;
  /**
   * Sigla da Unidade da Federação
   * xsd:TUfEmi
   */
  UF: TUfEmi;
  /**
   * Número do Documento de Arrecadação de Receita
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  nDAR?: string;
  /**
   * Data de emissão do DAR (AAAA-MM-DD)
   * xsd:TData
   */
  dEmi?: string;
  /**
   * Valor Total constante no DAR
   * xsd:TDec_1302, pattern `0|0\.[0-9]{2}|[1-9]{1}[0-9]{0,12}(\.[0-9]{2})?`
   */
  vDAR?: string;
  /**
   * Repartição Fiscal emitente
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  repEmi: string;
  /**
   * Data de pagamento do DAR (AAAA-MM-DD)
   * xsd:TData
   */
  dPag?: string;
};
// ...
export type TNFe_infNFe_dest = {
  /**
   * Razão Social ou nome do destinatário
   * xsd:TString, length 2..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  xNome?: string;
  /** Dados do endereço */
  enderDest?: TEndereco;
  /**
   * Indicador da IE do destinatário:
   * 1 – Contribuinte ICMSpagamento à vista;
   * 2 – Contribuinte isento de inscrição;
   * 9 – Não Contribuinte
   */
  indIEDest: "1" | "2" | "9";
  /**
   * Inscrição Estadual (obrigatório nas operações com contribuintes do ICMS)
   * xsd:TIeDestNaoIsento, length 0..14, pattern `[0-9]{2,14}`
   */
  IE?: string;
  /**
   * Inscrição na SUFRAMA (Obrigatório nas operações com as áreas com benefícios de incentivos fiscais sob controle da SUFRAMA) PL_005d - 11/08/09 - alterado para aceitar 8 ou 9 dígitos
   * pattern `[0-9]{8,9}`
   */
  ISUF?: string;
  /**
   * Inscrição Municipal do tomador do serviço
   * xsd:TString, length 1..15, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  IM?: string;
  /**
   * Informar o e-mail do destinatário. O campo pode ser utilizado para informar o e-mail
   * de recepção da NF-e indicada pelo destinatário
   * xsd:TString, length 1..60, pattern `[!-ÿ]{1}[ -ÿ]{0,}[!-ÿ]{1}|[!-ÿ]{1}`
   */
  email?: string;
} & (
  ({
  /**
   * Número do CNPJ
   * xsd:TCnpj, length 0..14, pattern `[0-9A-Z]{12}[0-9]{2}`
   */
  CNPJ: string;
  CPF?: never; idEstrangeiro?: never;
})
  | ({
  /**
   * Número do CPF
   * xsd:TCpf, length 0..11, pattern `[0-9]{11}`
   */
  CPF: string;
  CNPJ?: never; idEstrangeiro?: never;
})
  | ({
  /**
   * Identificador do destinatário, em caso de comprador estrangeiro
   * pattern `([!-ÿ]{0}|[!-ÿ]{5,20})?`
   */
  idEstrangeiro: string;
  CNPJ?: never; CPF?: never;
})
);
// ...
export type TNFe_infNFe_autXML = {

} & (
  ({
  /**
   * CNPJ Autorizado
   * xsd:TCnpj, length 0..14, pattern `[0-9A-Z]{12}[0-9]{2}`
   */
  CNPJ: string;
  CPF?: never;
})
  | ({
  /**
   * CPF Autorizado
   * xsd:TCpf, length 0..11, pattern `[0-9]{11}`
   */
  CPF: string;
  CNPJ?: never;
})
);
// ...
export type TNFe_infNFe = {
  /**
   * Versão do leiaute (v4.00)
   * @attribute xsd:TVerNFe, pattern `4\.00`
   */
  versao: string;
  /**
   * PL_005d - 11/08/09 - validação do Id
   * @attribute pattern `NFe[0-9]{6}[0-9A-Z]{12}[0-9]{26}`
   */
  Id: string;
  /** identificação da NF-e */
  ide: TNFe_infNFe_ide;
  /** Identificação do emitente */
  emit: TNFe_infNFe_emit;
  /** Emissão de avulsa, informar os dados do Fisco emitente */
  avulsa?: TNFe_infNFe_avulsa;
  /** Identificação do Destinatário */
  dest?: TNFe_infNFe_dest;
  /** Identificação do Local de Retirada (informar apenas quando for diferente do endereço do remetente) */
  retirada?: TLocal;
  /** Identificação do Local de Entrega (informar apenas quando for diferente do endereço do destinatário) */
  entrega?: TLocal;
  /**
   * Pessoas autorizadas para o download do XML da NF-e
   * occurs 0..10
   */
  autXML?: TNFe_infNFe_autXML[];
  /**
   * Dados dos detalhes da NF-e
   * occurs 1..990
   */
  det: TNFe_infNFe_det[];
  /** Dados dos totais da NF-e */
  total: TNFe_infNFe_total;
  /** Dados dos transportes da NF-e */
  transp: TNFe_infNFe_transp;
  /** Dados da cobrança da NF-e */
  cobr?: TNFe_infNFe_cobr;
  /** Dados de Pagamento. Obrigatório apenas para (NFC-e) NT 2012/004 */
  pag: TNFe_infNFe_pag;
  /** Grupo de Informações do Intermediador da Transação */
  infIntermed?: TNFe_infNFe_infIntermed;
  /** Informações adicionais da NF-e */
  infAdic?: TNFe_infNFe_infAdic;
  /** Informações de exportação */
  exporta?: TNFe_infNFe_exporta;
  /** Informações de compras  (Nota de Empenho, Pedido e Contrato) */
  compra?: TNFe_infNFe_compra;
  /** Informações de registro aquisições de cana */
  cana?: TNFe_infNFe_cana;
  /** Informações do Responsável Técnico pela emissão do DF-e */
  infRespTec?: TInfRespTec;
  /** Grupo para informações da solicitação da NFF */
  infSolicNFF?: TNFe_infNFe_infSolicNFF;
  /** Produtos Agropecurários Animais, Vegetais e Florestais */
  agropecuario?: TNFe_infNFe_agropecuario;
  /** Grupo de Informação do Provedor de Assinatura e Autorização */
  infPAA?: TNFe_infNFe_infPAA;
};
// ...
export const TNFe_infNFe_ide: CT<TNFe_infNFe_ide> = { id: "TNFe.infNFe.ide", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ e: "cUF", t: st_TCodUfIBGE }, { e: "cNF", t: st$1 }, { e: "natOp", t: st$2 }, { e: "mod", t: st_TMod }, { e: "serie", t: st_TSerie }, { e: "nNF", t: st_TNF }, { e: "dhEmi", t: st_TDateTimeUTC }, { e: "dhSaiEnt", t: st_TDateTimeUTC, n: 0 }, { e: "dPrevEntrega", t: st_TData, n: 0 }, { e: "tpNF", t: st$3 }, { e: "idDest", t: st$4 }, { e: "cMunFG", t: st_TCodMunIBGE }, { e: "cMunFGIBS", t: st_TCodMunIBGE, n: 0 }, { e: "tpImp", t: st$5 }, { e: "tpEmis", t: st$6 }, { e: "cDV", t: st$7 }, { e: "tpAmb", t: st_TAmb }, { e: "finNFe", t: st_TFinNFe }, { e: "tpNFDebito", t: st_TTpNFDebito, n: 0 }, { e: "tpNFCredito", t: st_TTpNFCredito, n: 0 }, { e: "indFinal", t: st$3 }, { e: "indPres", t: st$8 }, { e: "indIntermed", t: st$3, n: 0 }, { e: "cIndOp", t: st$9, n: 0 }, { e: "procEmi", t: st_TProcEmi }, { e: "verProc", t: st$10 }, { g: "s", i: [{ e: "dhCont", t: st_TDateTimeUTC }, { e: "xJust", t: st$11 }], n: 0 }, { e: "NFref", t: TNFe_infNFe_ide_NFref, n: 0, x: 999 }, { e: "gCompraGov", t: TCompraGov, n: 0 }, { e: "gPagAntecipado", t: TNFe_infNFe_ide_gPagAntecipado, n: 0 }] } };
// ...
export const TNFe_infNFe_emit: CT<TNFe_infNFe_emit> = { id: "TNFe.infNFe.emit", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ g: "c", i: [{ e: "CNPJ", t: st_TCnpj }, { e: "CPF", t: st_TCpf }] }, { e: "xNome", t: st$18 }, { e: "xFant", t: st$2, n: 0 }, { e: "enderEmit", t: TEnderEmi }, { e: "IE", t: st_TIe, n: 0 }, { e: "IEST", t: st_TIeST, n: 0 }, { g: "s", i: [{ e: "IM", t: st$22 }, { e: "CNAE", t: st$23, n: 0 }], n: 0 }, { e: "CRT", t: st$24 }, { e: "ISUFEmit", t: st$25, n: 0 }] } };
export const TNFe_infNFe_avulsa: CT<TNFe_infNFe_avulsa> = { id: "TNFe.infNFe.avulsa", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ e: "CNPJ", t: st_TCnpj }, { e: "xOrgao", t: st$2 }, { e: "matr", t: st$2 }, { e: "xAgente", t: st$2 }, { e: "fone", t: st$21, n: 0 }, { e: "UF", t: st_TUfEmi }, { e: "nDAR", t: st$2, n: 0 }, { e: "dEmi", t: st_TData, n: 0 }, { e: "vDAR", t: st_TDec_1302, n: 0 }, { e: "repEmi", t: st$2 }, { e: "dPag", t: st_TData, n: 0 }] } };
// ...
export const TNFe_infNFe_dest: CT<TNFe_infNFe_dest> = { id: "TNFe.infNFe.dest", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ g: "c", i: [{ e: "CNPJ", t: st_TCnpj }, { e: "CPF", t: st_TCpf }, { e: "idEstrangeiro", t: st$26 }] }, { e: "xNome", t: st$18, n: 0 }, { e: "enderDest", t: TEndereco, n: 0 }, { e: "indIEDest", t: st$28 }, { e: "IE", t: st_TIeDestNaoIsento, n: 0 }, { e: "ISUF", t: st$25, n: 0 }, { e: "IM", t: st$22, n: 0 }, { e: "email", t: st$2, n: 0 }] } };
export const TLocal: CT<TLocal> = { id: "TLocal", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ g: "c", i: [{ e: "CNPJ", t: st_TCnpjOpc }, { e: "CPF", t: st_TCpf }] }, { e: "xNome", t: st$18, n: 0 }, { e: "xLgr", t: st$18 }, { e: "nro", t: st$2 }, { e: "xCpl", t: st$2, n: 0 }, { e: "xBairro", t: st$18 }, { e: "cMun", t: st_TCodMunIBGE }, { e: "xMun", t: st$18 }, { e: "UF", t: st_TUf }, { e: "CEP", t: st$1, n: 0 }, { e: "cPais", t: st$27, n: 0 }, { e: "xPais", t: st$18, n: 0 }, { e: "fone", t: st$21, n: 0 }, { e: "email", t: st$2, n: 0 }, { e: "IE", t: st_TIe, n: 0 }] } };
export const TNFe_infNFe_autXML: CT<TNFe_infNFe_autXML> = { id: "TNFe.infNFe.autXML", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "c", i: [{ e: "CNPJ", t: st_TCnpj }, { e: "CPF", t: st_TCpf }] } };
// ...
export const TNFe_infNFe_total: CT<TNFe_infNFe_total> = { id: "TNFe.infNFe.total", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ e: "ICMSTot", t: TNFe_infNFe_total_ICMSTot }, { e: "ISSQNtot", t: TNFe_infNFe_total_ISSQNtot, n: 0 }, { e: "retTrib", t: TNFe_infNFe_total_retTrib, n: 0 }, { e: "ISTot", t: TISTot, n: 0 }, { e: "IBSCBSTot", t: TIBSCBSMonoTot, n: 0 }, { e: "vNFTot", t: st_TDec_1302, n: 0 }] } };
// ...
export const TNFe_infNFe: CT<TNFe_infNFe> = { id: "TNFe.infNFe", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "versao", t: st_TVerNFe, r: 1 }, { a: "Id", t: st$0, r: 1 }], c: { g: "s", i: [{ e: "ide", t: TNFe_infNFe_ide }, { e: "emit", t: TNFe_infNFe_emit }, { e: "avulsa", t: TNFe_infNFe_avulsa, n: 0 }, { e: "dest", t: TNFe_infNFe_dest, n: 0 }, { e: "retirada", t: TLocal, n: 0 }, { e: "entrega", t: TLocal, n: 0 }, { e: "autXML", t: TNFe_infNFe_autXML, n: 0, x: 10 }, { e: "det", t: TNFe_infNFe_det, x: 990 }, { e: "total", t: TNFe_infNFe_total }, { e: "transp", t: TNFe_infNFe_transp }, { e: "cobr", t: TNFe_infNFe_cobr, n: 0 }, { e: "pag", t: TNFe_infNFe_pag }, { e: "infIntermed", t: TNFe_infNFe_infIntermed, n: 0 }, { e: "infAdic", t: TNFe_infNFe_infAdic, n: 0 }, { e: "exporta", t: TNFe_infNFe_exporta, n: 0 }, { e: "compra", t: TNFe_infNFe_compra, n: 0 }, { e: "cana", t: TNFe_infNFe_cana, n: 0 }, { e: "infRespTec", t: TInfRespTec, n: 0 }, { e: "infSolicNFF", t: TNFe_infNFe_infSolicNFF, n: 0 }, { e: "agropecuario", t: TNFe_infNFe_agropecuario, n: 0 }, { e: "infPAA", t: TNFe_infNFe_infPAA, n: 0 }] } };